requirements.txt
sentence-transformers
scikit-learn
hdbscan
networkx
weaviate-client
pytest
